refactor: Extract test code templates for better maintainability
- Created separate template files for test code blocks
- Improved code organization and reusability
- Fixed test script for correct input order
- Added missing datetime import in export template
